+++ catalogue/CatalogueTest.cpp
@@ -231,6 +231,7 @@ namespace {
     tapeDrive.desiredForceDown = false;
     tapeDrive.diskSystemName = "dummyDiskSystemName";
     tapeDrive.reservedBytes = 694498291384;
+    tapeDrive.reservationSessionId = 0;
     return tapeDrive;
   }
 
@@ -245,6 +246,7 @@ namespace {
     tapeDrive.desiredForceDown = false;
     tapeDrive.diskSystemName = "dummyDiskSystemName";
     tapeDrive.reservedBytes = 694498291384;
+    tapeDrive.reservationSessionId = 0;
 
     tapeDrive.sessionStartTime = 1001;
     tapeDrive.mountStartTime = 1002;