+++ tapeserver/castor/tape/tapeserver/daemon/DataTransferSessionTest.cpp
@@ -295,6 +295,7 @@ public:
     tapeDrive.desiredForceDown = false;
     tapeDrive.diskSystemName = "dummyDiskSystemName";
     tapeDrive.reservedBytes = 694498291384;
+    tapeDrive.reservationSessionId = 0;
     return tapeDrive;
   }